consolidate global_parallel with global.
